requirements.txt
img2pdf==0.4.4
ocrmypdf==13.4.3
Pillow==9.1.0
pymongo==4.1.1
requests==2.27.1

pdf2image~=1.16.0

PyYAML~=6.0
PyPDF2~=1.27.12
docx2pdf~=0.1.8
